Revert "ANDROID: mmc: move to a SCHED_FIFO thread"
This reverts commit eea999d7487ac5b48ce48d4be60b7c7377aae7df.
As a part of above commit, mmcqd thread's priority was
modified to sched_fifo making it a RT task. This resulted
in starving other tasks like watchdog and starting hiting
watchdog bite error. This is expected if there are too
many requests to SD card and since it is fast enough to
serve the requests, it may starve other tasks.

Hence reverting the change to avoid such issues.

Change-Id: I620b674a990b9e2e76027a7419e1275679fa4eb0
Signed-off-by: Sayali Lokhande <sayalil@codeaurora.org>
Signed-off-by: Dušan Uverić <dusan.uveric9@gmail.com>