msm: kgsl: Write the profiling reference data before submission
Update the GPU ticks and CPU time in the user profiling buffer
before sending the command to the GPU. This removes a small
race condition where the GPU updates the execution profiling
data before the CPU has updated the reference data.
In this case the reference data is still set to the default
value when the profiling data is read by the user.

Change-Id: I65c9d5c3235eeee2353816d65d8eefdab961f9f8
Signed-off-by: Harshdeep Dhatt <hdhatt@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>
Signed-off-by: Stallix <team.evolutionx.rom@gmail.com>
Signed-off-by: Dušan Uverić <dusan.uveric9@gmail.com>