msm: kgsl: Use alignment corresponding to max supported page size
As 1MB is the max page size supported by kgsl, use the alignment
corresponding to 1MB page while selecting the page size for a
memory request. This will make sure that higher order pages are
used if possible for a given memory size and will also improve
the usage of higher order kgsl memory pools and general performance.

Change-Id: Ib4a5885a2e2c27c4a9e3f1589ede8290f9edbc9a
Signed-off-by: Deepak Kumar <dkumar@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>
Signed-off-by: Stallix <team.evolutionx.rom@gmail.com>
Signed-off-by: Dušan Uverić <dusan.uveric9@gmail.com>