Revert "vidc: do not bias/boost cpu/tasks during video instances"
For now, there isn't a way to distinguish between picture in picture scenarios
and other video instances without the help of userspace. Also, picture in picture
does not invoke a close and re-open of a new vidc instance, so I cannot invoke
disable_schedtune_boost() from there either.

So, I'll just drop this to prevent a drop in UX performance in those cases.

This reverts commit 11498d72a66877453cd5ee21119de7710ceac0d2.

Signed-off-by: idkwhoiam322 <idkwhoiam322@raphielgang.org>
Signed-off-by: nysascape <jago@nysascape.digital>
Signed-off-by: Dušan Uverić <dusan.uveric9@gmail.com>